Transfer onboarding course lists to adapter dsl
